Moved camp logics to AI from app-level managers.
