+++ src/engine/core/ai/camp/CampManager.test.ts
