+++ src/engine/core/ai/camp/index.ts
@@ -1,0 +1,3 @@
+export * from "@/engine/core/ai/camp/camp_types";
+export * from "@/engine/core/ai/camp/CampConfig";
+export * from "@/engine/core/ai/camp/CampManager";
