+++ src/engine/core/database/zones.ts
@@ -1,7 +1,7 @@
+import { CampManager } from "@/engine/core/ai/camp";
 import { IRegistryObjectState } from "@/engine/core/database/database_types";
 import { registerObject, resetObject, unregisterObject } from "@/engine/core/database/objects";
 import { registry } from "@/engine/core/database/registry";
-import { CampManager } from "@/engine/core/managers/camp";
 import { GameObject, Optional, Vector } from "@/engine/lib/types";
 
 /**
