--- src/engine/core/managers/camp/index.ts
@@ -1,3 +1,0 @@
-export * from "@/engine/core/managers/camp/camp_types";
-export * from "@/engine/core/managers/camp/CampConfig";
-export * from "@/engine/core/managers/camp/CampManager";
