+++ src/engine/core/schemes/stalker/walker/actions/ActionWalkerActivity.ts
@@ -1,10 +1,10 @@
 import { action_base, LuabindClass } from "xray16";
 
+import { CampManager } from "@/engine/core/ai/camp/CampManager";
 import { StalkerPatrolManager } from "@/engine/core/ai/patrol/StalkerPatrolManager";
 import { animpoint_predicates } from "@/engine/core/animation/predicates/animpoint_predicates";
 import { EStalkerState } from "@/engine/core/animation/types";
 import { getCampZoneForPosition, registry, setStalkerState } from "@/engine/core/database";
-import { CampManager } from "@/engine/core/managers/camp/CampManager";
 import { GlobalSoundManager } from "@/engine/core/managers/sounds/GlobalSoundManager";
 import { IAnimpointActionDescriptor } from "@/engine/core/schemes/stalker/animpoint/animpoint_types";
 import { ISchemeWalkerState } from "@/engine/core/schemes/stalker/walker";
